# Conflicts:
#	core/src/main/java/lucee/runtime/tag/Http.java

+++ core/src/main/java/lucee/runtime/type/QueryImpl.java
@@ -363,7 +363,7 @@ public class QueryImpl implements Query,Objects {
 	private void setAttributes(Statement stat,int maxrow, int fetchsize,TimeSpan timeout) throws SQLException {
 		if(maxrow>-1) stat.setMaxRows(maxrow);
         if(fetchsize>0)stat.setFetchSize(fetchsize);       
-        if(timeout!=null && timeout.getSeconds()>0)
+        if(timeout!=null && ((int)timeout.getSeconds())>0)
         	DataSourceUtil.setQueryTimeoutSilent(stat,(int)timeout.getSeconds());
 	}
